Rename ybadge to yboard
